Fixed hanging state machine based on Bridge's feedback.
Also the transfer pneumatic was changed so that whenever the robot is in the final "climb" state of the climb state machine it no longer changes the status of the pneumatic based on the bound controller button for it.  This is so that the pickup is always up during the final stage of the climb.